Fix the semaphore timeout
Was a int32 overflow since our timeout values were 32 bit variable and
we were multiplying them with a large value, the result was a 32 bit
value also. Just changing the timeout constants to 64 bit values fixes
that.
I still think this timeout value implementation is a bit stupid, why
not have the current enums (eAckTimeout…) be the timeout values instead
of being an index into another array which has the real values?